backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
joblib
requests
python-dotenv
xgboost